Initial version of forms and views for mailing list submissions.
 The nonwg_lists/submit wizard is nearly done; it needs templates
and to figure out why submit from step 2 goes back to step 1 instead
of calling done.

 The req_list wizard is just started.  It needs to get the incoming
value of domain_name as an alternat version of dname, and be able to
get dname from GET args to handle the javascript redirect.  It also
needs all the workflow ("I'm allowed to create this list",
who-is-approver, etc.) and additional forms.



git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@122 7b24d068-2d4e-4fce-9bd7-cbd2762980b0