Use get() instead of filter()[0].  This results in a better exception on failure (DoesNotExist instead of IndexError).
Call the database query once and store it in a temp variable instead
of doing the same query twice.

Handle day_id=NULL for the morning beverages/breakfast time.

Use the DoesNotExist exception instead of a generic except: for
the area - except: is likely to hide things that you didn't expect.



git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@156 7b24d068-2d4e-4fce-9bd7-cbd2762980b0