Create "allononeline" filter and use it for list_wizard_done_email.
Use the Site to get the URL in list_wizard_done_email.

Set req for list_wizard_done_email as it is for approval email.

Only use domain-based extra approvers for non-WG requests.

Supply hidden fields for mlist_name and domain_name for deletion.
(These should be visible but uneditable, but passing them
through is a start)


git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@267 7b24d068-2d4e-4fce-9bd7-cbd2762980b0