Merged [2958] from shane@time-travellers.org: Updated the JavaScript associated with the search page to control whether the 'search' button is enabled or not. It is now enabled if either the main search box has input, or if any of the sub-options have input, or both.
Note that the Yahoo! widgets we were using did weird things 
to the button name, so I disabled that. This changes the
look and feel of the button - that may be a contenious issue.
OTOH, I think that this is probably a good thing, since this
is the only place where we override the user's default 
appearance settings.

Another possible issue is the use of JavaScript without
JQuery. The existing code here did not use JQuery, so I 
decided to simply extend it.

See the Trac ticket #480 for more.

http://trac.tools.ietf.org/tools/ietfdb/ticket/480 

git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@2966 7b24d068-2d4e-4fce-9bd7-cbd2762980b0