Merged in [8293] from rjsparks@nostrum.com:\n    XXX: Should what is now captured in the list of STATUSCHANGE_RELATIONS instead be captured in the database in name_docrelationshipname?
    Refactored editing the notify field to remove redundant code.
Changed the default notification list to include .all for documents, and the wg list for wg documents.
Allowed recalculating the notification list for all document types.
Improved the calculated notification list value for charters, conflict-reviews, and status-changes.
Adds shepherds to the notification list when they are assigned to a document.
Adds the working group email list to the notification list when a document is adopted.
Fixes #1438 

git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@8385 7b24d068-2d4e-4fce-9bd7-cbd2762980b0