Merged in [10770], [10771], [10772], and [10773] from housley@vigilsec.com:
The IESG Telechat Agenda now shows "(Has RFC Editor Note)" after the I-D
filename if there is an RFC Editor Note associated with the document.  This
was added to the html and txt versions of the agenda.  It was not added to the
Secretariat view or the Scribe view of the agenda.

For transition, when an AD edits the RFC Editor Note, they need to move the
text from the current writeup into the new field.  Returning documents on the
telechat agenda seems to be the biggest opportunity for something to fall
between the cracks.  If an event of type "changed_rfc_editor_note' exists, and
the string "RFC Editor Note" appears in the text of the most recent
'changed_ballot_writeup_text' event, then a message is shown that tells the AD
to remove the RFC Editor Note from the writeup.



git-svn-id: https://svn.tools.ietf.org/svn/tools/ietfdb/trunk@10783 7b24d068-2d4e-4fce-9bd7-cbd2762980b0